pyproject.toml
[tool.poetry]
name = "linotak"
version = "0.1.0"
description = "Link Note Taker (a microblog app that emphasizes short notes linking to other sites)."
authors = ["Damian Cugley <pdc@alleged.org.uk>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
django = "^4.2.8"
django-environ = "^0.11.0"
celery = ">=4.4.0,<4.5.0"
requests = "^2.28.1"
psycopg2 = "^2.9.5"
Markdown = "^3.4.1"
requests-oauthlib = "^1.3.1"
urllib3 = "2.2.3"
uritemplate = "^4.1.1"

[tool.poetry.group.dev.dependencies]
tblib = "^1.7.0"
factory-boy = "^3.2.1"
httpretty = "^1.1.4"
flake8 = "^6.0.0"
black = "^22.10.0"
flake8-black = "^0.3.5"
pep8-naming = "^0.13.2"
isort = "^6.0.1"

[tool.isort]
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
